pyproject.toml
[tool.poetry]
name = "django-url-checks"
version = "0.1.0"
description = "Django checks to ensure valid URL Configurations"
keywords = ["django", "check", "url", "uri", "slash"]
authors = ["Andrew Pinkham <code@andrewsforge.com>"]
license = "BSD-2-Clause"
readme = "README.md"
repository = "https://github.com/jambonsw/django-url-checks"
packages = [
    { include = "url_checks", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.6"
pytest-cov = "^2.6"

[tool.poetry.dev-dependencies]
pytest = "^4.0"
pre-commit = "^1.12"
tox = "^3.6"
pytest-django = "^3.4"

[tool.black]
line-length = 79
py36 = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
